Responded to comments
